refactor: extract config loading in separate package
