+++ cmd/migrate.go
@@ -2,6 +2,7 @@ package cmd
 
 import (
 	"github.com/odpf/siren/app"
+	"github.com/odpf/siren/config"
 	"github.com/spf13/cobra"
 )
 
@@ -14,6 +15,6 @@ func init() {
 }
 
 func migrate(cmd *cobra.Command, args []string) error {
-	c := app.LoadConfig()
+	c := config.LoadConfig()
 	return app.RunMigrations(c)
 }
