+++ config/config.go
@@ -1,4 +1,4 @@
-package app
+package config
 
 import (
 	"fmt"
